Use test grouping for MongoDB extensions
Fixes #3089
