+++ integration-test-groups/mongodb/pom.xml
@@ -22,26 +22,23 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-build-parent-it</artifactId>
+        <artifactId>camel-quarkus-integration-test-groups</artifactId>
         <version>2.2.0-SNAPSHOT</version>
-        <relativePath>../poms/build-parent-it/pom.xml</relativePath>
     </parent>
 
-    <artifactId>camel-quarkus-integration-test-groups</artifactId>
+    <artifactId>camel-quarkus-integration-tests-mongodb</artifactId>
     <packaging>pom</packaging>
 
-    <name>Camel Quarkus :: Integration Test Groups</name>
+    <name>Camel Quarkus :: MongoDB Integration Tests</name>
 
     <properties>
         <quarkus.banner.enabled>false</quarkus.banner.enabled>
     </properties>
 
     <modules>
-        <!-- test groups a..z; do not remove this comment, it is important when sorting via  mvn process-resources -Pformat -->
-        <module>aws2</module>
-        <!-- <module>azure</module> disabled by cq-prod-maven-plugin:prod-excludes -->
-        <module>foundation</module>
+        <!-- extensions a..z; do not remove this comment, it is important when sorting via  mvn process-resources -Pformat -->
         <module>mongodb</module>
+        <!-- <module>mongodb-gridfs</module> disabled by cq-prod-maven-plugin:prod-excludes -->
     </modules>
 
 </project>
