+++ integration-tests-support/mongodb/src/main/resources/initMongodb.txt
