+++ poms/bom-test/pom.xml
@@ -114,6 +114,11 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.camel.quarkus</groupId>
+                <artifactId>camel-quarkus-integration-tests-support-mongodb</artifactId>
+                <version>${camel-quarkus.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel.quarkus</groupId>
                 <artifactId>camel-quarkus-integration-wiremock-support</artifactId>
                 <version>${camel-quarkus.version}</version>
             </dependency>
